pyproject.toml
[project]
name = "ringrl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aim>=3.29.1",
    "gymnasium[atari,box2d,classic-control,other]>=1.2.3",
    "numpy>=2.4.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "torch>=2.9.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.12",
]

[tool.mypy]
python_version = "3.12"
show_error_codes = true
pretty = true
warn_unused_ignores = true

# Ignore virtualenv + run artifacts
exclude = '(^\.venv/|^\.aim/|^checkpoints/)'

[[tool.mypy.overrides]]
module = ["yaml.*"]
ignore_missing_imports = true

